+++ projects/net.splitcells.dem.api/src/main/java/net/splitcells/dem/data/set/SetFactoryImplConfigured.java
@@ -22,23 +22,23 @@ import net.splitcells.dem.lang.annotations.JavaLegacyBody;
 
 import java.util.Optional;
 
-import static net.splitcells.dem.data.set.SetFI_deterministic.setFI_deterministic;
-import static net.splitcells.dem.data.set.SetFI_random.setFI_random;
+import static net.splitcells.dem.data.set.SetFactoryImplDeterministic.setFactoryImplDeterministic;
+import static net.splitcells.dem.data.set.SetFactoryImplRandom.setFactoryImplRandom;
 import static net.splitcells.dem.data.set.SetLegacyWrapper.setLegacyWrapper;
 
-public class SetFI_configured implements SetF {
+public class SetFactoryImplConfigured implements SetF {
     public static SetF setFiConfigured() {
-        return new SetFI_configured();
+        return new SetFactoryImplConfigured();
     }
 
     private SetF setF;
 
-    private SetFI_configured() {
+    private SetFactoryImplConfigured() {
         final var isDeterministic = Dem.configValue(IsDeterministic.class);
         if (isDeterministic.isPresent() && isDeterministic.get().isTrue()) {
-            setF = setFI_deterministic();
+            setF = setFactoryImplDeterministic();
         } else {
-            setF = setFI_random();
+            setF = setFactoryImplRandom();
         }
     }
 
@@ -60,11 +60,11 @@ public class SetFI_configured implements SetF {
     @Deprecated
     private void update(Optional<Bool> oldValue, Optional<Bool> newValue) {
         if (newValue.isEmpty()) {
-            setF = new SetFI();
+            setF = new SetFactoryImpl();
         } else if (newValue.get().isTrue()) {
-            setF = setFI_deterministic();
+            setF = setFactoryImplDeterministic();
         } else {
-            setF = setFI_random();
+            setF = setFactoryImplRandom();
         }
     }